Main Header: Get events from user context rather than flux call
